Update `fromRequestHeaders` to allow specifying a list of request headers
This can be used to only set specific request headers, such as `User-Agent`.

+++ headers.php
@@ -156,13 +156,28 @@ class Headers implements HeadersInterface
 		}
 	}
 
-	public static function fromRequestHeaders():? HeadersInterface
+	public static function fromRequestHeaders(string ...$include):? HeadersInterface
 	{
 		if (function_exists('getallheaders')) {
-			$headers = new self(getallheaders());
-			$headers->delete('host');
-			$headers->delete('cookie');
-			return $headers;
+			if (count($include) === 0) {
+				$headers = new self(getallheaders());
+				$headers->delete('host');
+				$headers->delete('cookie');
+
+				return $headers;
+			} else {
+				$include = array_map('strtolower', $include);
+				$headers = new self();
+
+				foreach (getallheaders() as $key => $value) {
+					$key = strtolower($key);
+					if (in_array($key, $include)) {
+						$headers->set($key, $value);
+					}
+				}
+
+				return $headers;
+			}
 		} else {
 			return null;
 		}
